rework csshelper.ts
moved var to theme, so we only need to change it there and we can also access the single value (e.g. if you only need padding-left)

Added makeResponivePadding and makeResponiveMargin for paddings that actually need to change on larger screens (content paddings will stay the same for now)
Thought about putting it in a prob inside `makeMargin`but that way we would have a lot more TS overhead.